+++ packages/workeros-programs/src/index.js
@@ -96,4 +96,7 @@ export const libraries = [
   { path: "/lib/workeros-node/require-runtime.js", source: () => fetchText("./node/require-runtime.js") },
   { path: "/lib/workeros-node/fs.js", source: () => fetchText("./node/fs.js") },
   { path: "/lib/workeros-node/path.js", source: () => fetchText("./node/path.js") },
+  { path: "/lib/workeros-node/os.js", source: () => fetchText("./node/os.js") },
+  { path: "/lib/workeros-node/url.js", source: () => fetchText("./node/url.js") },
+  { path: "/lib/workeros-node/module.js", source: () => fetchText("./node/module.js") },
 ];